make sublimetext3/python3 friendly
ST3 changes the python path to make the current directory a python
module. This means all of the modules bundled with the plugin are
submodules of 'isort', and combined with the absolute imports, it
causes a lot of problems while trying to import the plugin.

To fix this, the ImportErrors are caught, and imports are tried
again using their absolute paths